requirements.txt
stem
psutil>=2.2.1
netifaces>=0.10.4
pyvirtualdisplay
selenium>=2.45.0
-e git+https://github.com/webfp/tor-browser-selenium.git#egg=tbselenium-1.0
